Bug 12820 - SIP2 does not send rental charge fee!
Koha's SIP2 implementation has facilities to send the fee amount ( BV )
for checking out a given item, but in our code, the fee amount is never
set!

It's even more complicated than that. If there is a fee, the sip server
checks to see if BO ( fee acknowledged ) is set to Y. If it is not, the
SIP server is supposed to block the checkout and send a screen message
with the fee amount. If it is, the checkout proceeds as usual.

Test Plan:
1) Apply the patch for bug 13159
2) Set a checkout fee for an itemtype
3) Find an item of the itemtype
4) Run the SIP cli utility:
   sip_cli_emulator.pl -a <address> -p <port> -su <sip_user> -sp <sip_user_password> -l <location> --patron <patron_cardnumber> -m checkout --item <item_barcode>
5) Note the server returns an AF field describing the fee and doesn't
   check out the item.
6) Rerun step 4 with the addition of the -fa flag
7) Note the item checks out to the patron this time